signed/unsigned mismatch
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@2887 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/ResourceEditor.cpp
@@ -77,7 +77,7 @@ CResourceEditor::CResourceEditor(BYTE* pbPE, int iSize) {
     }
 
     // Invalid section pointer (goes beyond the PE image)
-    if (sectionHeadersArray[i].PointerToRawData > m_iSize)
+    if (sectionHeadersArray[i].PointerToRawData > (unsigned int)m_iSize)
       throw runtime_error("Invalid section pointer");
   }
